Rename collections.py to metrics_collections.py (#695)
